Extract QMvec
