+++ src/common/calc_mesh.cc
@@ -2,6 +2,7 @@
 // Implementation of Jacobian methods for script interfaces
 // ========================================================================
 
+#include "toastdef.h"
 #include "calc_mesh.h"
 
 void BuildMesh(QMMesh **mesh, int nvtx, int nel, int dim, int nnd0,
